Rename time-helper.js
